requirements.txt
requests
feedparser
notion-client
python-dotenv
google-generativeai
langdetect